add category queries

+++ app/Models/Direction.php
@@ -52,7 +52,7 @@ class Direction extends Model
             if(!$image->get()->isEmpty()) unlink(public_path($image->pluck('src')->first()));
 
             $icon = $direction->icon;
-            unlink(public_path($icon));
+            if($icon) unlink(public_path($icon));
 
             $direction->images()->delete();
             $direction->keywords()->delete();
